rename and moved Client stuff to right folder
