+++ src/main/java/com/johanneslosch/recall/client/Recall.java
@@ -1,4 +1,4 @@
-package com.johanneslosch.recall;
+package com.johanneslosch.recall.client;
 
 import com.johanneslosch.recall.util.ConfigReader;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
@@ -12,6 +12,7 @@ public class Recall  extends TelegramLongPollingBot {
             //message content
             update.getMessage().getText();
             System.out.println(String.format("RECEIVED: message: %s -> at: %s -> from: %s", update.getMessage().getText(), System.nanoTime(), update.getMessage().getChat().getUserName()));
+            Client.client.send(String.format("RECEIVED: message: %s -> at: %s -> from: %s", update.getMessage().getText(), System.nanoTime(), update.getMessage().getChat().getUserName()));
         }
     }
